refactor and bullet mode

+++ build/socket/disconnect/index.js
@@ -8,10 +8,20 @@ const events_1 = require("../../mock/events");
 const disconnect = (deps, socket, roomQueue, io) => {
     const roomService = new room_1.default(deps);
     return async () => {
+        const roomList = await roomService.list();
         const room = await roomService.disconnectPlayer(socket.id);
         if (!room)
             return;
-        console.log('Disconnect');
+        const currentRoom = roomList.find((room) => {
+            return room.playerList.some((player) => {
+                return player.socketId === socket.id;
+            });
+        });
+        if (!currentRoom)
+            return;
+        const user = currentRoom.playerList.find((player) => {
+            return player.socketId === socket.id;
+        });
         if (!room.playerList.length) {
             await roomService.delete(room._id);
             roomQueue.removeQueue(`catching_play_${room.hash}`);
@@ -19,7 +29,7 @@ const disconnect = (deps, socket, roomQueue, io) => {
             roomQueue.removeQueue(`joinRoom_${room.hash}`);
         }
         await socket.leave(room.hash);
-        await socket.broadcast.to(room.hash).emit(events_1.USER.DISCONNECT, room);
+        await socket.broadcast.to(room.hash).emit(events_1.USER.DISCONNECT, { room, user });
     };
 };
 exports.default = disconnect;
